refactor: 将 `CertificateMap` 移动到 `package cert`
