+++ core/client_test.go
@@ -19,7 +19,7 @@ import (
 	"github.com/wechatpay-apiv3/wechatpay-go/core/auth"
 	"github.com/wechatpay-apiv3/wechatpay-go/core/auth/signers"
 	"github.com/wechatpay-apiv3/wechatpay-go/core/auth/verifiers"
-	"github.com/wechatpay-apiv3/wechatpay-go/core/cert/certificate_map"
+	"github.com/wechatpay-apiv3/wechatpay-go/core/cert"
 	"github.com/wechatpay-apiv3/wechatpay-go/core/consts"
 	"github.com/wechatpay-apiv3/wechatpay-go/utils"
 
@@ -117,7 +117,7 @@ func TestClient_Upload(t *testing.T) {
 	}
 
 	verifier = verifiers.NewSHA256WithRSAVerifier(
-		certificate_map.NewCertificateMap(
+		cert.NewCertificateMap(
 			map[string]*x509.Certificate{testWechatCertSerialNumber: wechatPayCertificate},
 		),
 	)
